requirements.txt
einops==0.5.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
numpy==1.23.4
pillow==9.0.1
pytorch-fid==0.3.0
scikit-learn==1.2.2
scipy==1.10.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tqdm==4.64.1
transformers==4.23.1